Use Message Collector for checks
