+++ test/hw_tests/hwtest_scan_compare.test
@@ -1,7 +1,5 @@
 <launch>
-  <arg name="continuous_run" default="false"/>
-  <arg name="number_of_days" default="1"/>
-  <arg name="time_limit" value="$(eval number_of_days*86400 if continuous_run else 10)"/>
+  <arg name="test_duration" default="10"/>
 
   <!--<include file="$(find psen_scan_v2)/launch/psen_scan_v2.launch" />-->
 
@@ -11,8 +9,8 @@
     <remap from="/laser_scanner/scan" to="/laser_scanner/scan_reference"/>
   </node>
 
-  <test test-name="scan_compare" pkg="psen_scan_v2" type="hwtest_scan_compare" time-limit="$(arg time_limit)">
+  <test test-name="scan_compare" pkg="psen_scan_v2" type="hwtest_scan_compare" time-limit="$(eval test_duration + 10)">
     <param name="testfile" value="$(find psen_scan_v2)/test/hw_tests/template.bag"/>
-    <param name="continuous_run" value="$(arg continuous_run)"/>
+    <param name="test_duration" value="$(arg test_duration)"/>
   </test>
 </launch>
